danga distributed lock daemon
take 1

based mostly on mogilefs code, which was based largely on perlbal
code, both of which use Danga::Socket.  so writing this was cake.

currently it has commands for:

* trylock(lockname)  tries to get lock, returns error immediately
      if couldn't get it.  clients can take as many locks as they want

* releaselock(lockname)   but locks are automatically unlocked
       when client disconnects

* locks()  shows existing locks.  mostly for debugging.

* noop()    just to keep connection alive, if clients want it.  not
            actually necessary, since on Linux uses tcp keep-alives


NOTE:  all this is useless without a good multi-lock-daemon-aware client.
that's coming next.  (don't want the lock daemon being a single POF)

this will eventually be used for all LJ's distributed locking.


git-svn-id: http://code.sixapart.com/svn/ddlockd/trunk@2 75bc8fc0-0210-0410-8e5e-a32055405bc5